I fix links for zora drops using points metadata.
