+++ lib/zora/getCollectionUrl.tsx
@@ -1,6 +1,9 @@
 import { Address } from 'viem'
+import { baseSepolia } from 'viem/chains'
 
-const getCollectionUrl = (collectionAddress: Address) =>
-  `https://testnet.zora.co/collect/bsep:${collectionAddress}/1`
+const getCollectionUrl = (chainId: any, collectionAddress: Address) => {
+  const isTestnet = chainId == baseSepolia.id
+  return `https://${isTestnet ? 'testnet.' : ''}zora.co/collect/${isTestnet ? 'bsep' : 'base'}:${collectionAddress}/1`
+}
 
 export default getCollectionUrl
